+++ CHANGELOG.md
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The keybinds shown in the volume creation tool are now shortened. (<kbd>Control</kbd> becomes <kbd>Ctrl</kbd>, and the spaces around <kbd>+</kbd> are removed)
 - Made the copy keybindings use [Amecs](https://modrinth.com/project/rcLriA4v) instead of being hardcoded to use <kbd>Control</kbd>. This does mean **Entity Selector Tools depends on Amecs** for now. I do want to make it optional though, so I'm still working on that.
 - Changed the color of inline code in the config screen to gold from light purple.
+- If Axiom is loaded, its mouse icons are used instead of the name of the keybind.
+- *Box Creation as Axiom Tool* is now enabled by default.
+- Rewrite of compatibility with Axiom 
 - Several miscellaneous code improvements and renames.
 
 ## [1.0.0]